+++ python/desc/slrealizer/__init__.py
@@ -4,3 +4,5 @@ from deblend import *
 from testing import *
 from test import *
 from moment import *
+from null_deblend_v1 import *
+from null_deblend_v2 import *
