+++ python/desc/slrealizer/null_deblend_v1.py
@@ -1,0 +1,111 @@
+# ======================================================================                                    
+from __future__ import print_function
+from scipy.stats import multivariate_normal
+import plotting
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib.ticker import NullFormatter, MaxNLocator
+from numpy import linspace
+import matplotlib
+import scipy
+import scipy.optimize as opt
+import math
+import photutils
+from astropy.visualization import SqrtStretch
+from astropy.visualization.mpl_normalize import ImageNormalize
+from photutils import CircularAperture
+from photutils import DAOStarFinder
+import moment
+import desc.slrealizer
+import scipy.stats
+from scipy.stats import moment
+from scipy import ndimage
+import skimage
+import skimage.measure
+from skimage.measure import moments
+from moment import covariance_matrix
+# ======================================================================                                  
+
+"""                                                                                                    
+Given a specific date and the OM10 catalog, this deblends the sources that are on the catalog.           
+Assumes null deblender where all the sources are assumed to be observed as single objects.                 
+All the sources are assumed to have Gaussian PSFs.                                                 
+ """
+#global variable that controls the size of the plot                                                       
+x_min = -5.0
+x_max = 5.0
+y_min = -5.0
+y_max = 5.0
+distance = 0.01
+
+number_of_rows = int((x_max - x_min)/distance)
+number_of_columns = int((y_max - y_min)/distance)
+
+
+#global variable that controls the size of the plot
+x_min = -5.0
+x_max = 5.0
+y_min = -5.0
+y_max = 5.0
+distance = 0.01
+
+def null_deblend_v1(data):
+    x, y = np.mgrid[x_min:x_max:distance, y_min:y_max:distance]
+    pos = np.dstack((x, y))
+    xbar, ybar, covariance_matrix = intertial_axis(data)
+    zeroth_moment = data.sum()
+    first_moment_x = x_min + (xbar) * distance
+    first_moment_y = y_min + (ybar) * distance
+    rv = scipy.stats.multivariate_normal([first_moment_x,first_moment_y], covariance_matrix, allow_singular=True) #FIX BUG           
+    image = [[0]*number_of_rows for _ in range(number_of_columns)]
+    image = image + rv.pdf(pos)*zeroth_moment
+    print('**************zeroth moment: ', zeroth_moment)
+    print('**************first moment: ', first_moment_x, first_moment_y)
+    print('**************second moment: ', covariance_matrix)
+    return image
+
+
+
+
+def please_work(data):
+    xbar, ybar, cov = intertial_axis(data)
+    fig, ax = plt.subplots()
+    ax.imshow(data)
+    plot_bars(xbar, ybar, cov, ax)
+    plt.show()
+
+def raw_moment(data, iord, jord):
+    nrows, ncols = data.shape
+    y, x = np.mgrid[:nrows, :ncols]
+    data = data * x**iord * y**jord
+    return data.sum()
+
+def intertial_axis(data):
+    """Calculate the x-mean, y-mean, and cov matrix of an image."""
+    data_sum = data.sum()
+    m10 = raw_moment(data, 1, 0)
+    m01 = raw_moment(data, 0, 1)
+    x_bar = m10 / data_sum
+    y_bar = m01 / data_sum
+    u11 = (raw_moment(data, 1, 1) - x_bar * m01) / data_sum
+    u20 = (raw_moment(data, 2, 0) - x_bar * m10) / data_sum
+    u02 = (raw_moment(data, 0, 2) - y_bar * m01) / data_sum
+    cov = np.array([[u20, u11], [u11, u02]])
+    return x_bar, y_bar, cov
+
+def plot_bars(x_bar, y_bar, cov, ax):
+    """Plot bars with a length of 2 stddev along the principal axes."""
+    def make_lines(eigvals, eigvecs, mean, i):
+        """Make lines a length of 2 stddev."""
+        std = np.sqrt(eigvals[i])
+        print('standard deviation is: ', std)
+        vec = 2 * std * eigvecs[:,i] / np.hypot(*eigvecs[:,i])
+        x, y = np.vstack((mean-vec, mean, mean+vec)).T
+        return x, y
+    print("This is the covariance I calculated: ", cov)
+    print("This is the eigenvalue I have", np.linalg.eigh(cov))
+    mean = np.array([x_bar, y_bar])
+    eigvals, eigvecs = np.linalg.eigh(cov)
+    ax.plot(*make_lines(eigvals, eigvecs, mean, 0), marker='o', color='white')
+    ax.plot(*make_lines(eigvals, eigvecs, mean, -1), marker='o', color='red')
+    ax.axis('image')
